refactor: remove business logic, keep auth-only base
- Remove all Application business logic except Auth (login/register/refresh + access)
- Remove all WebAPI controllers except Auth
- Fix Direction.cs and DirectionConfiguration.cs syntax (duplicated trailing code)
- Fix StaffRoleType enum comparison in Thesis repositories (remove int cast)
- Remove runtime seeding from ApplicationDbContextInitialiser (keep migrations only)
- Remove SeedAsync call from Program.cs
- Add docs/seed_edu_data.sql for idempotent Edu_* table seeding

Total: 488 files changed, 24332 deletions

+++ src/AWM.Service.Infrastructure/Persistence/Repositories/Thesis/DirectionRepository.cs
@@ -60,7 +60,7 @@ public sealed class DirectionRepository : IDirectionRepository
         return await _context.StaffAssignments
             .AsNoTracking()
             .Where(a => a.UserId == userId &&
-                        a.RoleType == (int)StaffRoleType.Supervisor &&
+                        a.RoleType == StaffRoleType.Supervisor &&
                         a.TargetEntityType == "Direction" &&
                         a.IsActive && !a.IsDeleted)
             .Join(_context.Directions.Where(d => !d.IsDeleted && d.SemesterId == academicYearId),
